requirements.txt

aiosignal==1.3.2
attrs==25.3.0
discord-py==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.3.2
pip==25.0
python-dotenv==1.1.0
yt-dlp==2025.3.31